requirements.txt
certifi==2020.12.5
MouseInfo==0.1.3
Pillow==8.0.1
PyAutoGUI==0.9.52
pygame==2.0.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
PyTweening==1.0.3

tqdm==4.54.1
wincertstore==0.2

opencv-python~=4.4.0.46
numpy~=1.19.3
pytesseract~=0.3.7
screeninfo~=0.8.1

Flask==2.1.0
psycopg2-binary==2.9.1